backend/python/upload/requirements.txt
# Core Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0

# File processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-magic>=0.4.27

# Cloud storage
boto3>=1.26.0
django-storages>=1.13.0

# Background tasks
celery>=5.3.0
flower>=2.0.0

# Security and moderation
pyclamd>=0.5.0

# Video processing
ffmpeg-python>=0.2.0

# Image processing and filters
scikit-image>=0.20.0

# File validation
python-magic-bin>=0.4.14

# Development and testing
pytest>=7.2.0
pytest-django>=4.5.0
coverage>=7.2.0
black>=23.3.0
flake8>=6.0.0

# Monitoring and logging
sentry-sdk>=1.25.0
django-extensions>=3.2.0
